requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beartype==0.21.0
beautifulsoup4==4.13.4
blis==1.3.0
browserbase==1.4.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
fastai==2.8.2
fastcore==1.8.2
fastdownload==0.0.7
fastprogress==1.0.3
fasttransform==0.0.2
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
google-auth==2.40.3
google-genai==1.19.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
networkx==3.5
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.2.1
playwright==1.52.0
plum-dispatch==2.5.7
preshed==3.0.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
rejex==0.0.1.1
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
wrapt==1.17.2
